1、删除专栏的时候如果有教程，则不删除，教程的 ID 改为统一的 Long 型；3、查询专栏的时候，如果为空则不再查询用户

+++ paicoding-service/src/main/java/com/github/paicoding/forum/service/article/service/ColumnSettingService.java
@@ -45,14 +45,14 @@ public interface ColumnSettingService {
      *
      * @param columnId
      */
-    void deleteColumn(Integer columnId);
+    void deleteColumn(Long columnId);
 
     /**
      * 删除专栏文章
      *
      * @param id
      */
-    void deleteColumnArticle(Integer id);
+    void deleteColumnArticle(Long id);
 
     /**
      * 通过关键词，从标题中找出相似的进行推荐，只返回主键 + 标题
